requirements.txt
# Core Libraries
RPi.GPIO>=0.7.0
smbus2>=0.4.2
adafruit-circuitpython-vl53l0x>=5.1.1
pybluez>=0.23
pyserial>=3.5

# Sensor Libraries
mpu6050-raspberrypi>=1.2.0
adafruit-blinka>=8.0.0

# System Monitoring
psutil>=5.9.0

# Logging
loguru>=0.7.0

# Configuration
pyyaml>=6.0.1

# Optional: For more advanced GPIO control
pigpio>=1.78






# Core Python Dependencies
python-can==4.1.0
RPi.GPIO==0.7.1
smbus2==0.4.2

# Sensor Libraries
adafruit-circuitpython-vl53l0x==5.1.10
mpu6050-raspberrypi==1.2.0

# Communication Protocols
pyserial==3.5
bluepy==1.3.0

# Data Processing and Utilities
numpy==1.22.4
scipy==1.8.1

# Logging and Configuration
python-json-logger==2.0.4
pyyaml==6.0

# Testing Frameworks
pytest==7.3.1
mock==5.0.2
unittest2==1.1.0

# Optional: Performance Monitoring
psutil==5.9.5

# Development and Debugging Tools
ipython==8.13.2
pylint==2.17.4